fix: fixed dropdown component structure
