+++ src/common/components/inputs/index.ts
@@ -1,0 +1,2 @@
+export { default as Input } from "./input";
+export { default as InputDropdown } from "./input-dropdown";
