+++ src/const/sex.ts
@@ -1,0 +1,10 @@
+export const sex = [
+  {
+    value: "Male",
+    label: "Male",
+  },
+  {
+    value: "Female",
+    label: "Female",
+  },
+];
